server/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
google-generativeai==0.4.1
